third_party/blink/web_tests/external/wpt/content-security-policy/style-src/inline-style-allowed-while-cloning-objects.sub-expected.txt

This is a testharness.js-based test.
[FAIL] inline-style-allowed-while-cloning-objects 18
  assert_equals: expected "rgb(238, 130, 238) none repeat scroll 0% 0% / auto padding-box border-box" but got "rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box"
[FAIL] inline-style-allowed-while-cloning-objects 19
  assert_equals: expected "rgb(238, 130, 238) none repeat scroll 0% 0% / auto padding-box border-box" but got "rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box"
[FAIL] non-HTML namespace
  assert_equals: expected "" but got "background: rgb(238, 130, 238)"
Harness: the test ran to completion.
